+++ Engine/Engine/Object.h
@@ -22,7 +22,7 @@ public:
 	~Object();
 
 	virtual void Init(Renderer* renderer, std::string fileName, DirectX::XMFLOAT4 color);
-	void Tick(double dt) override;
+	void Tick(float dt) override;
 	void Render(Renderer* renderer);
